+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.23] 2024-01-05
+
+### Added
+
+- Now **CGameUserSettings** supports so called console variables. A console variable is an option (integer, float or string) that could be defined through two main approaches:
+using in-game console or defined in settings.cfg file.
+
 ## [0.6.23] 2023-12-30
 
 ### Added